Ensure height and width syncs with drawing_data
Before that, drawing_data was updated only when normal data updates.
This change also adds separate validation step in __init__ because if we
assign to self.width before self.height, data couldn't be updated
because there is no self.height. Thus the data has to be modified
directly in __init__.